pyproject.toml
[project]
name = "contour"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fal-client>=0.11.0",
    "fastapi>=0.128.0",
    "google-genai>=1.57.0",
    "numpy>=2.4.0",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.21",
    "rasterio>=1.5.0",
    "uvicorn>=0.40.0",
]
